apply suggestions
