+++ cmd/util/util.go
@@ -3,7 +3,6 @@
 package util
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -14,7 +13,6 @@ import (
 	"github.com/openfga/openfga/pkg/storage/postgres"
 	"github.com/openfga/openfga/pkg/storage/sqlcommon"
 	storagefixtures "github.com/openfga/openfga/pkg/testfixtures/storage"
-	"github.com/openfga/openfga/pkg/typesystem"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/require"
@@ -88,83 +86,4 @@ func PrepareTempConfigFile(t *testing.T, config string) {
 	_, err = confFile.WriteString(config)
 	require.Nil(t, err)
 	require.Nil(t, confFile.Close())
-}
-
-type ValidationResult struct {
-	StoreID       string
-	ModelID       string
-	IsLatestModel bool
-	Error         string
-}
-
-// ValidateAllAuthorizationModels lists all stores and then, for each store, lists all models.
-// Then it runs validation on each model.
-func ValidateAllAuthorizationModels(ctx context.Context, db storage.OpenFGADatastore) ([]ValidationResult, error) {
-	validationResults := make([]ValidationResult, 0)
-
-	continuationTokenStores := ""
-
-	for {
-		// fetch a page of stores
-		stores, tokenStores, err := db.ListStores(ctx, storage.PaginationOptions{
-			PageSize: 100,
-			From:     continuationTokenStores,
-		})
-		if err != nil {
-			return nil, fmt.Errorf("error reading stores: %w", err)
-		}
-
-		// validate each store
-		for _, store := range stores {
-
-			latestModelID, err := db.FindLatestAuthorizationModelID(ctx, store.Id)
-			if err != nil {
-				fmt.Printf("no models in store %s \n", store.Id)
-			}
-
-			continuationTokenModels := ""
-
-			for {
-				// fetch a page of models for that store
-				models, tokenModels, err := db.ReadAuthorizationModels(ctx, store.Id, storage.PaginationOptions{
-					PageSize: 100,
-					From:     continuationTokenModels,
-				})
-				if err != nil {
-					return nil, fmt.Errorf("error reading authorization models: %w", err)
-				}
-
-				// validate each model
-				for _, model := range models {
-					_, err := typesystem.NewAndValidate(model)
-
-					validationResult := ValidationResult{
-						StoreID:       store.Id,
-						ModelID:       model.Id,
-						IsLatestModel: model.Id == latestModelID,
-					}
-
-					if err != nil {
-						validationResult.Error = err.Error()
-					}
-					validationResults = append(validationResults, validationResult)
-				}
-
-				continuationTokenModels = string(tokenModels)
-
-				if continuationTokenModels == "" {
-					break
-				}
-			}
-		}
-
-		// next page of stores
-		continuationTokenStores = string(tokenStores)
-
-		if continuationTokenStores == "" {
-			break
-		}
-	}
-
-	return validationResults, nil
 }
