+++ cmd/validate_models.go
@@ -10,6 +10,7 @@ import (
 	"github.com/openfga/openfga/pkg/storage/mysql"
 	"github.com/openfga/openfga/pkg/storage/postgres"
 	"github.com/openfga/openfga/pkg/storage/sqlcommon"
+	"github.com/openfga/openfga/pkg/typesystem"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -38,6 +39,13 @@ func NewValidateCommand() *cobra.Command {
 	return cmd
 }
 
+type ValidationResult struct {
+	StoreID       string `json:"store_id"`
+	ModelID       string `json:"model_id"`
+	IsLatestModel bool   `json:"is_latest_model"`
+	Error         string `json:"error"`
+}
+
 func runValidate(_ *cobra.Command, _ []string) error {
 	engine := viper.GetString(datastoreEngineFlag)
 	uri := viper.GetString(datastoreURIFlag)
@@ -58,14 +66,14 @@ func runValidate(_ *cobra.Command, _ []string) error {
 	case "memory":
 		fallthrough
 	default:
-		return fmt.Errorf("invalid datastore engine type: %s", engine)
+		return fmt.Errorf("storage engine '%s' is unsupported", engine)
 	}
 
 	if err != nil {
-		return fmt.Errorf("could not connect to database")
+		return fmt.Errorf("failed to open a connection to the datastore: %v", err)
 	}
 
-	validationResults, err := util.ValidateAllAuthorizationModels(ctx, db)
+	validationResults, err := ValidateAllAuthorizationModels(ctx, db)
 	if err != nil {
 		return err
 	}
@@ -78,4 +86,76 @@ func runValidate(_ *cobra.Command, _ []string) error {
 	fmt.Println(string(marshalled))
 
 	return nil
+}
+
+// ValidateAllAuthorizationModels lists all stores and then, for each store, lists all models.
+// Then it runs validation on each model.
+func ValidateAllAuthorizationModels(ctx context.Context, db storage.OpenFGADatastore) ([]ValidationResult, error) {
+	validationResults := make([]ValidationResult, 0)
+
+	continuationTokenStores := ""
+
+	for {
+		// fetch a page of stores
+		stores, tokenStores, err := db.ListStores(ctx, storage.PaginationOptions{
+			PageSize: 100,
+			From:     continuationTokenStores,
+		})
+		if err != nil {
+			return nil, fmt.Errorf("error reading stores: %w", err)
+		}
+
+		// validate each store
+		for _, store := range stores {
+
+			latestModelID, err := db.FindLatestAuthorizationModelID(ctx, store.Id)
+			if err != nil {
+				fmt.Printf("no models in store %s \n", store.Id)
+			}
+
+			continuationTokenModels := ""
+
+			for {
+				// fetch a page of models for that store
+				models, tokenModels, err := db.ReadAuthorizationModels(ctx, store.Id, storage.PaginationOptions{
+					PageSize: 100,
+					From:     continuationTokenModels,
+				})
+				if err != nil {
+					return nil, fmt.Errorf("error reading authorization models: %w", err)
+				}
+
+				// validate each model
+				for _, model := range models {
+					_, err := typesystem.NewAndValidate(model)
+
+					validationResult := ValidationResult{
+						StoreID:       store.Id,
+						ModelID:       model.Id,
+						IsLatestModel: model.Id == latestModelID,
+					}
+
+					if err != nil {
+						validationResult.Error = err.Error()
+					}
+					validationResults = append(validationResults, validationResult)
+				}
+
+				continuationTokenModels = string(tokenModels)
+
+				if continuationTokenModels == "" {
+					break
+				}
+			}
+		}
+
+		// next page of stores
+		continuationTokenStores = string(tokenStores)
+
+		if continuationTokenStores == "" {
+			break
+		}
+	}
+
+	return validationResults, nil
 }
